requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
alembic
python-jose
passlib[bcrypt]
python-multipart
pydantic-settings